pyproject.toml
[tool.poetry]
name = "graphql-tracing"
version = "0.1.0"
description = ""
authors = ["Mitchell Friedman <mitchfriedman5@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.0.3"
ariadne = "^0.14.1"
honeycomb-beeline = "3.1.0"
SQLAlchemy = "^1.4.31"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
